LayoutTests/fast/css/outline-auto-zoom-canvas-rotated-scale-expected.txt

This test ensures that the canvas drawFocusIfNeeded focus ring scales correctly under a rotated non-uniform transform, exercising the singular value decomposition.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pixelsRotatedScale > pixelsIdentity is true
PASS successfullyParsed is true

TEST COMPLETE
